requirements.txt
requests==2.24.0
beautifulsoup4==4.9.3
python-telegram-bot==13.6
firebase-admin==5.0.0
pytz==2021.1
